+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 販売者編集エンドポイントの汎用性強化
+- /events/individualScreeningEvent を /events/screeningEvent のエイリアスとして非推奨に設定
 
 ### Deprecated